CA: change meaning of "scrambled" flag in meta.
The rationale is this, before, the scrambled flag
would be set if ecm's are recorded. This means that
recording would still be handled as "scrambled" even
though it already had been descrambled and ecm's were
only recorded as backup. Usually we don't want
descrambling of recordings already descrambled,
because it will request ecm's even though they're
not used, possibly at high speed, possibly locking
the card. The drawback is that if you want to
descramble a recording that was partly descrambled
(salvage it), you will need to adjust this flag
manually (using a text editor), set it to '1'.