CA: Use descrambling demux when we descramble -> allocate demux from start.
This solves an issue for services that are recognised as "pvr"
(fe == nullptr) and need to have a descrambling demux but don't get
one. At this point enigma assumes recordings and streams are always
descrambled (which is obviously not the case (anymore)).

In practice this was only an issue on a very limited set of receivers
where some of the demuxers are non-descrambling.

(cherry picked from commit 0bfbb1e2148c6073cc98e58774d91098ee797460)
Signed-off-by: Erik Slagter <erik@openpli.org>