NumberZap: Auto adjust NumberZap timeout time
In the past the initial key had a timeout of 3 seconds. When you press the
second key the timeout was reduced by 1 seconds.

But it seems there are slow button pressers between us where 1 sec was to
fast,

Now the auto OK after the second key is adjusted to the time the first two
buttons are pressed. Let see if this solves the issue as described here:

http://forums.openpli.org/topic/40260-channel-number-zap-timeout/page-2

When this does not solve that specific described issue we should consider
to revert this commit and add a config (fast/medium/slow) instead.

I pushed this commit because I prefer to have as less as configs as
possible and I would give it a chance. Here it is working as expected the
complete afternoon.