FBC: unlink FBC leaf tuner when it's closed.
So it can be linked again at some point, to another
FBC root tuner, or the same.

Also move the code out of the header file (ugly...)

This commit will need additional commits to not set FBC leaf
tuners to "loop through" by default, but to "not connected"
instead, and only then, the full potential of the FBC tuners
can be used.