FrontendInfo: Finetune tuner string options
Add value for number of tuners with spaces and max tuners. Spaces can be
removed to allow more room and when there are to manu tuners the non used
tuners can be removed. Default room for 6 tuners with spaces and 10
without defined. (e.g. in my FullNight skin I need to reserve a bit more
spaces to display all 10 tuners of the solo4K)