bsod: Dump a plain text log file instead of XML
Since the log file is to be read and interpreted by humans, and since
XML is hard to read for both humans and machines, just output the
crash log in a plain text file which is easy to parse for both.

It also gets rid of a lot of string mangling and XML voodoo, reducing
the amount of code and memory allocations while already crashing.
It won't make another two in-memory copies of the whole log buffer, for
example.

Remove the attempt to blame others using maintainer.info because this
never really worked anyway.