eDVBFrontend: do not check for if FE_CAN_MULTISTREAM defined
The FE_CAN_MULTISTREAM is enumeration not a defined value, so it is not possible to use preprocessor to check if defined.
Compiling with older libc is possible using a small compatibility header with all missing definitions.

(cherry picked from commit 2a9d110880f056336d5d296cd6613d844eff43d8)
Signed-off-by: Erik Slagter <erik@openpli.org>