Use FullNightHD skin by default if your box supports a full HD skin
Also block settings to full HD when your box do not support it.

I hope I cached all boxes that do not support a full 1920x1080 skin as far
I know only Zgemma and XP1000 are effected. Please notify me if there are
more as they need to be added in the SystemInfo.py file