Do not auto default to FullNight-HD 1920x1080 skin
As a lot of users do not understand what overscan means and/or do not
understand how to setup their TV correctly we received a lot of questions
about this new skin. Therefore I "reverted" the auto switch to the full HD
1920x1080 more modern skin for now.

It might be considered later to revert this commit as soon the users do understand when they configured their TV not correctly and/or it is understand how to select a non in the overscan writing skin when their TV does not support it.