Add repeat wakeup messages for HDMI CEC
In some cases when e.g. a surround set is connected via HDMI between the
receiver and the TV it could be the TV is in cascade turned on via the
surround send, then it might happen the TV is missing some HDMI-CEC
commands. While repeating the wakeup comments when the TV is turned on we
ensure the TV will receive these commands