Complete reconstructure of MenuSort
Default does not change anything
A-Z simply does alphabetic order
When on 'user' in a menu you can press the blue button (not indicated in
skin). Then with yellow or OK you can hide/show items and with the <> next
to 0 button you can move items. With Blue button you can reset all custom
configurations. With Red or Cancel you Abort and with Green button you
save the changes.