Add IP-addresses of networks to About screen
Todo: There is a wish to also add the IPV6 addresses, however we need to
find a way to retreive these addressen in an easy way. E.g. by considering
to add netifaces library to the image -or- instead of the socket parse
subroutine we need to parse 'ifconfig' via the console instead.