NimManager: Change multiType initialization order
Change multiType initialization before start configuring the frontend, because the multiType
configured in settings might be different than the one parsed from nim_sockets.

Before this was not a problem because we always consider multiType can be C or T, but when
multituner device is used, then we are not configuring correctly the S (eg advanced).