atsc, eit: skip events from the past
It turns out EIT-0 is not always kept up-to-date, it often contains an
old event, which we should skip.
Unfortunately we have only 3 hours of events in EIT-0, so when the
information is outdated, we might end up with only one event (now), the
'next' event will be missing in that case.
Or we might even end up with no events at all.
(a solution would be to parse EIT-1 as well, but that is too much of a
hassle for now, and we can always fallback on the epgcache)