Don't touch avahi-daemon
Enigma2 starts and stops the avahi deamon for no apparent reason. The daemon
is smart enough to listen to network interface registration and does not need
Enigma2 to babysit it.

Since I plan to use avahi inside E2, messing with the daemon doesn't look
like a proper thing to do.