Solve issue where fbc tuners were not correctly found
Previously to indicate a non unicable configuration was tricked by using -1 as value for SatCR_idx. Since the JESS introduction we have a nice SatCR_config that indicates 0 when it is non Unicable or Jess, 1 for Unicable and 2 for Jess. In fbc there was still a 'left over' in the code that checks on SatCR_idx. This commit changes this test correctly to SatCR_format.

This solves also an issue that since 30 December the leave tuners of the Solo4K were suddenly not available due to this corrective commit: https://github.com/OpenPLi/enigma2/commit/27239517769d980e13074dff17d55c19296f079a

For more detailed information read the discussion (in Dutch) here: https://forums.openpli.org/topic/47106-tuner-problemen-solo-4k-met-image-van-vandaag-en-eerder/page-2