Return back duplicate translation in the config
This can be useful after you change the language in the start wizard.
It does not provide a full translation for all.
It works only in one direction- from the English to other language.
But it is very likely that in the start wizard you need to translate only the config values after language change from English to your local.