Postpone resolving default recording path, and allow to revert it
Once a movie recording path had been selected, there was no way of
reverting that choice. Offer to select the default location as a
valid option, causing the system to pick the largest local disk
again.

This also postpones the selection of the default location to the
moment it's actually used, so that it really picks the larger internal
storage rather than what was biggest on startup.