streamserver: introduce stopStreamClient function
The stopStreamClient can be used to stop specific stream
by passing ip and serviceref.

TODO: Above function should be integrated into StreamClientInfo,
where  streaming clients can be selected and stopped.

A possible improvement with a small modification might be the pass
of ip or serviceref only, in order to stop all clients from the same ip
or all clients on the same serviceref.