CableScan: Only use connected DVB-C nims
It was always the fact that all DVB-C tuners were selectable in the cable
scan even when they are not connected. Since we introduce automatic tuner
selection for the cable scan it also could select such not connected tuner
leading to a side effect that the cablescan was not working anymore a
nothing could be tuned. This commit does remove the not connected tuners
from the nimlist used for the Cablescan

Read the discussion for more information https://forums.openpli.org/topic/49247-kabelscan-werkt-niet-meer/page-8