dvbci: revert testing on ci device nodes to access().
It appeared some manufacturers device drivers export some ci
device nodes, so they can be find in a directory listing, but
userspace isn't not allowed to open them. That's where using
open instead of access makes a difference. Now reverting to the
previous behaviour.

This should fix the problems for at least some of the users that
reported them the last few days.