Arrange center DVB graphical SUBS per channel
As there is no way to detect of they are already be centered. So now by default they are not centered and displayed as indicated by the provider. In case they are left alligned on a service and you want to center them you can set a flag via the channelselector