dvbci: support DVR as ci source
When a channel is not associated with a frontend, it must be a pvr
playback service. In that case, find the corresponding dvr device id 'n'
and configure the ci source to be DVR[n].

The second part of the patch makes sure the default ci source (tuner
"A") is restored, when the service is released.
This allows the ci device to release its dvr-specific resources, in case
the service was a pvr playback service.