CableScan/FastScan: Solve GSOD when remove/install these plugins.
WHen removing these plugins during a session it may lead
to GSOD when putting the box in standby when the autoscan function does
not exist anymore or when installing these plugins during a session you
can also get a GSOD when the Session global is None, so not initated.

Thanks for reporting Dimitrij.