ScanSetup.py: Multistream bug fix...
self.scan_sat.is_id is using "0 - 255" as the "limits" to ConfigInteger and "-1" as the "default" value. So the default is out of bounds. This means that once changed it is impossible for the user to return to the default value. This leads to various problems, from not being able to acquire tuner lock when scanning, to writing junk to lamedb. Also showing "-1" as a default value is meaningless to the user. How are they supposed to know this means NO_STREAM_ID FILTER? Zero cannot be used here as the default because that is a valid input stream id.

NO_STREAM_ID_FILTER is a CPP default. It would be good if this variable were available to Python in a similar way to "eDVBFrontendParametersSatellite.PLS_Root" rather than use a hard coded "-1" everywhere that some coders might not understand.

Please see the discussion on using a ConfigSelection field to deal with such a case: https://forums.openpli.org/topic/50339-configlist-item-advice/