Remove 1 minute from the goto deepstandby when in standby timer config
As does interfere with autoscan (cable and fastscan) and later also with
offline decoding of decoded recordings. In addition changing it to 5
minutes as minimum does not save that much energie. E.g. 4 minutes
longer with standby power of 20 W takes 750 times of using this timeout,
say 750 days, ~2 years, before we use 1 KWh... ~20 Eurocents