InputDevice: use correct IOC_SIZEBITS for non mips arch
Hardcoding the IOC_SIZEBITS is very bad because arch can override that value.
Use 13 for mips or the default 14.

Most probably it is be better to export that function from enigma2 to python.
So python will not have to "guess" the right values.

Above fixes the following error on arm:
[iInputDevices] getInputDevices event0 <ERROR: ioctl(EVIOCGNAME): [Errno 22] Invalid argument >