Get rid of ext3 to ext4 conversion tool
It's pretty hard to detect if a disk is ext3 (because the ext4 driver
mounts it, it appears in mtab as "ext4"), and it's been ages so if
people haven't run the conversion, they'll not be doing it anyway.

Remove the tool now that it outlived its usefulness so we no longer
have to answer difficult questions about it.