Flashimage: Solve issue when mounted drive is performed via a softlink
When you use the network browser to mount external drives it creates a
softlink in /media to the mount in /media/mnt. This shoftlink was used
as it was a hardware not mounted drive which creates the backup on that
device, which was later not available. This is solved ba (re)addint the
not os.path.islink check

Also avoid a GSOD when a directory is given by the backup plugin that
does not exists.

In develop this is not an issue as it seem there we have the islink
check