ChoiceBox: fix autoResize
For screen alignment, use relative counting depending on the screen resolution.
Removes count from hardcored 720x576 resolution, which allows aligment on all resolutions.

Remove widget text resize. We use the text size in skin as the output data for the screen size, so resize the text size is not required.
This is only the second time sets the existing size.