VideoHardware: use 'auto' aspect ratio policy when it is available
Components/AVSwitch.py adds the 'auto' option to the aspect ratio policy
setting choices when /proc/stb/video/policy(2)_choices contains 'auto'.

So when 'auto' is selected by the user, we should write 'auto' to the
hardware in order to select automatic aspect ratio policy mode.

Until now, the option 'bestfit' was written to the hardware, when 'auto'
was selected, the 'auto' hardware option was never used.
However, the 'bestfit' hardware option is already available through the
'scale' policy setting, which clearly is not the same as 'auto' aspect
ratio policy.