AVSwitch/VideoHardware: add support for 'full' aspect ratio mode
This option is similar to the 'scale' option (which maps to the
'bestfit' hardware setting).
Assume that on platforms where the 'full' option is available, the
'scale' (or: 'bestfit') option will try to fill as much of the screen as
possible, keeping the correct aspect ratio, while the 'fill' option will
always fill the entire screen.

Of course the implementation might differ between platforms. We just try
to make all available hardware settings available to the user.