[ActionMap] HelpableActionMaps multiple contexts
Add HelpableNumberActionMap to OpenPLi to match OpenViX and Beyonwiz images.

Allow HelpableActionMap and HelpableNumberActionMap to accept a list of contexts as well as a single context.  This makes it possible to convert ActionMap and NumberActionMap objects to helpable without changing their basic structure.  Single-context helpable ActionMaps are retained for backwards compatibility.

Simplify the ActionMap logging, print only one line instead of two for keymap errors.

PEP8 clean-up and fixes.

(This change will be required if the improved Beyonwiz HELP system is ever adopted.)

Most of the code written by Prl for the Beyonwiz image (Nov-2014 - Jan-2015)