[Setup.py] Cache and optimise loading of setup.xml
Cache the setup XML data in memory on the first access.  Every subsequent access will check if the source file has been changed.  If so, the file will be reloaded and reprocessed.  If not, the cached results are returned.

All setup menu titles are also extracted and cached to improve menu building time.  This code uses a derivative of the OpenViX "titleshort" attribute here known as "menuTitle" to allow Menu.py based menus to have alternate title text.

The "raise SetupError()" in "getSetupTitle()" has been retained as requested.  I still believe that this code should not crash just because a menu entry is missing.  It is enough to display an error message and log the error.  If the retention decision is reversed then remove the "SetupError()" class, the "raise SetupError()" line and associated comments.

This code change improves Setup menu and index load times.