[Setup.py] Clean up and optimise setup menu load
Optimise, improve and enhance generation of Setup menu item lists.  For OpenViX and Beyonwiz this change allows the list to be processed once for display rather than the two or three cycles in the previous versions.

Remove remnants of the setup.xml "separation" attribute directive.  This feature is now offered via the skin parameter "ConfigListSeperator".

Create a log entry if the Setup menu has no valid entries.

Obey Setup menu sort setting.

Correctly detect a Setup Level change and update the list accordingly.

If "%s %s" is found in an item's "text" or "description" attribute then replace it with the current box machine brand and model name.