[Setup.py] Move description code, add footnote
Move the description code from the SetupSummary class into the main Setup class where it is used.

Add Beyonwiz and OpenViX footnote support code.  An "*" at the end of a menu item displays a message to the user that changing this setting will result in a GUI Restart.

Add OpenViX menu path code to display the menu path to the current menu.