[Setup.py] Code clean for readability and tweaks
Reorganise the imports and code blocks to improve readability of the newly refactored code.

Check that the initialisation call back code is only added if not already added.

The "setup_key" skin facility appears to have been unused.  The name has been changed to "Setup_key" to better match existing skin names.  This option allows skin designers to create custom Setup based screens for any Setup menu item.

Rename the GREEN button text from "OK" to a more appropriate "Save".

The overall effect of this series of changes is to make the Setup.py features a superset of all builds while also adding new features and options.  The setp.xml file is only loaded and processed once per boot or GUI restart.  The source file is checked for any changes and the file will be reloaded should the file be changed.  This means that the GUI no longer has to be restarted to allow changes to setup.xml to be processed (OpenPLi).  All the changes and enhancements come with an up to four-fold reduction in time to generate any Setup.py based setup menu (Beyonwiz and OpenViX).  The time taken to build any setup menu lists (non Beyonwiz builds) is slightly longer for the first menu item but then about 50 time faster for all subsequent searches.