ConfigList: try fix GSOD on ShowHelp and HideHelp
Create functions onDeselect and onSelect in ConfigList to use it when call or return from VirtualKeyBoardCallback or MessageBox.
This is the realization of this Huevos idea: https://github.com/Taapat/enigma2/commit/bfb0b008115677c0a50c408b1b3373dde2002877 but in a safer way, and I hope it does not call GSOD.
Uses standard functions that are used in selectionChanged, and therefore they should work not only for HelpWindow, but also for other possible features.