[ActionMap] HelpableActionMaps multiple contexts
- Use a slightly longer form of initialisation for the ActionMap "__init__".  This format avoids an issue where the initialisation parameters become persistent across all invocations of this code.  This is inappropriate and undesirable.

In essence the problem with the original code is that it appears to be an efficient way to create a default empty list and dictionary.  What actually happens is that these items get created as a *persistent* objects.  Every invocation of the “__init__” method that doesn't specify the parameters will use that *SAME* (persistent) objects.  Any changes to either object will *persist* (be remembered) and be carried forward into every other invocation of this method!  THIS IS NOT WHAT SHOULD HAPPEN!

Please refer to http://pylint-messages.wikidot.com/messages:w0102 and https://nedbatchelder.com/blog/200806/pylint.html for more details.

- The HelpableActionMaps can now accept a list of contexts as well as supporting the original single context mode.  This should simplify adding help to existing ActionMaps. 

- ActionMap / keymap errors are now reported in a single log line rather than taking two lines.

- PEP8 clean-up and fixes.

These changes are fully backward compatible with all exiting Enigma2 code.  No changes to any existing code should be required required.

Most of the code written by Prl (Nov-2014 - Jan-2015).