Add multibootsupport for GigaBlue
Thanks and credits to TwolDE2 and also zgemma-star for the additional
info

It also simplifies some multiboot/flashimage related code
-but-
Somehow it is still a bit lame. It would be more helpfull when the
manufacturers did put all possible STARTUP files in some kind of format
to the startup partition so we simply can browse these files for
possebilities and simply copy a specific possible startup file to the
'master' startup file...e.g. having some STARTUP files in some kind of
format as STARTUP_BOOTSLOT_1_MODE_12 which we can parse.
Then we also do not have to calculate the 'parameters' as they are
simply in the possible slots and mode and when modi are added the
manufacturer adds a startup possebility.