[ActionMap.py] Revert multiple contexts component
This commit reverts the multiple contexts enhancement for the HelpableActionMap and HelpableNumberActionMap.

This enhancement caused an issue with a required legacy plugin (CSFD which is similar to IMDb but for https://www.csfd.cz/).  This plugin is closed source.  The author is not responding to calls for assistance.

It is possible that other plugins may also be negatively impacted so the enhancement will be reverted until such time it is safe to use for ALL plugins.