Multiboot: Arrange that the mmcblk0p1 is mounted to /tmp/startupmount
This ensures the STARTUP file can always written even when it is
unmounted to /media by the user (configuration). It also arranges that
it is mounted for Giga Blue devices.

2DO: It does not work for GB yet as the copying of the STARTUP slot
files to the STARTUP file must still be arranged. Hopefully TwolDE2 is
willing to help and test there. Please copy the files simply via
shutil.copyfile