Fix and Improve ActionMap (#1483)
* [ActionMap.py] Improve ActionMap initialisation

Use a slightly longer form of initialisation for the ActionMap `__init__`.  This format avoids an issue where the initialisation parameters become persistent across all invocations of this code.  This is inappropriate and undesirable.

In essence the problem with the original code is that it appears to be an efficient way to create a default empty list and dictionary.  What actually happens is that these items get created as a *persistent* objects.  Every invocation of the `__init__` method that doesn't specify the parameters will use that *SAME* (persistent) objects.  Any changes to either object will *persist* (be remembered) and be carried forward into every other invocation of this method!  THIS IS NOT WHAT SHOULD HAPPEN!

Please refer to http://pylint-messages.wikidot.com/messages:w0102 and https://nedbatchelder.com/blog/200806/pylint.html for more details.

* [ActionMap.py] Improve logging messages

ActionMap / keymap errors are now reported in a single log line rather than taking two lines.

* [ActionMap.py] Add HelpableNumberActionMap

Add a HelpableNumberActionMap for action map symmetry.

* [ActionMap.py] Improve HelpableActionMap initialisation

Use a slightly longer form of initialisation for the HelpableActionMap `__init__`.  This format avoids an issue where the initialisation parameters become persistent across all invocations of this code.  This is inappropriate and undesirable.

In essence the problem with the original code is that it appears to be an efficient way to create a default empty dictionary.  What actually happens is that this item gets created as a *persistent* object.  Every invocation of the `__init__` method that doesn't specify the parameter will use that *SAME* (persistent) object.  Any changes to the object will *persist* (be remembered) and be carried forward into every other invocation of this method!  THIS IS NOT WHAT SHOULD HAPPEN!

Please refer to http://pylint-messages.wikidot.com/messages:w0102 and https://nedbatchelder.com/blog/200806/pylint.html for more details.

* [ActionMap.py] PEP8 code tidy up

Perform a PEP8 clean up of the code and comments.