ActionMap: Remove not required work-a-round
Why put in a work-a-round for a non issue. It is in the code there from
2005 and it never harms. So why making the code here more complicated
then required.
I also do think the dangerous issue as described here http://pylint-messages.wikidot.com/messages:w0102 and https://nedbatchelder.com/blog/200806/pylint.html are not effective anymore -or- at least not in scope here, If they are still in scope we need to make a lot of changes in E2s python code.

When I really 'miss the boat' here please let us know so we consider to
revert this commit

In addition when e.g. the prameters are mandatory fields (always a need
here) we could better not give them a default at all...