Update from fr.po
Hi, I have reworked the translation from english to french for all messages in fr.po.
I don't use the word "service" in french because it is not used in french, we usually use "chaîne", "canal" or "transpondeur". 
I have a dreambox DM7080HD and I would be interested to do the same work for this receiver. My OS version is 4.2.11r7 (2016-05-11)  and my installed image of Enigma2 is dated from 2014-10-14. 
Can you help me??
Can you send to me the fr.po of this receiver?
In advance thank you for your help