NimManager: Add parameter to split combined jess/unicable switches/LNBs
There exist LNBs and switches that have on the lower SCRs unicable and
on higher SRCs jess. These were called dSCR in the unicable.xml. It did
fixed put the first 4 SRCs to unicable. Now we can add a parameter in
unicable.xml e.g. "dSRC,8" where the split is done on the first 8 SCRs
being unicable and the others jess.