VirtualKeyBoard: introduce wide buttons
This allows the use of buttons with a width of several buttons.
For example for space, shift or enter.
It allows you to create a symmetric keyboard that is similar to a computer keyboard.
You also can use for command buttons not only upperr case letters to make them look better.
The wide button is actually built from several individual buttons:
   LongL - left long button;
   LongM - midle long button;
   LongR - right long button.
If you want to use this you need to use a separator '__' in the keyboard before the button code to indicate the button type.
You can also hide the text by adding '__HIDE' after button code.
For example:
   u'LongL__Enter', u'LongR__ENTER__HIDE'
   u'LongL__SPACE__HIDE', u'LongM__SPACE__HIDE', u'LongM__SPACE__HIDE', u'LongR__SPACE__HIDE'

For example or test, you can use the layout of the Latvian keyboard, which will be changed after this commit.