Update fi.po (merge from openvix)
This commit merges the current 5+ years old Finnish translation with the latest translation file from the OpenVIX project. The Finnish translation in the OpenVIX project is far superior, both in terms of coverage and translation accuracy. Only the strings and translations common between the projects are taken, everything else remains from the previous translation.