VirtualKeyBoard: add the possibility to change the green button text
For example, when I use the plugin EPGSearch, I enter text in the virtual keyboard and press the green button to start searching instead of entering text somewhere.
This allow in this plugin and in others instead of Enter specify the text that would better describe what will be done after pressing the green or OK button.