Hotkey: By default assign GMEPG on long info/epg
This is helpfull for boxes that have only one epg/info (e.g. VU) as most
users want to have quick access. Now you need to 'adapt' it via hotkey
to have access to the GMEPG for one button epg/info boxes/remotes.
This change has one disadvantage, the showEventInfoPlugins is not
available without re-assign it to a different key and therefore I
defaulty assing this function by default to help-long.