Solve issue with KEY_TV in Radio mode with perm timeshift enabled
When permanent timeshift is enabled the KEY_TV is assign to go to live.
This function is also assigned to KEY_STOP. Due to this when in
permanent timeshift when we are in radio mode we cannot get back to TV
in the standard infobar due to this. By removing the key-assignement
KEY_TV in timeshift going back to TV in standard infobar will be
functional again. The function is still available via the STOP button
which as far I know each remote control should have (except for the
small DMM remote), In addition goto live TV is always still available
somehow by 'jumping' forward through the 'live-poínt'