[Directories.py] Revert "#" separator workaround
This patch resulted from an incorrect understanding of the actual problem.  The real issue was that some important images were specified to be in a shared location that is not in the normal search list.  I believe that the images do not exist in the nominated location.

The original issue can be corrected by placing the images at the specified location or else allowing the normal "Directories.py" logic to find the images in one of the standard locations.