[enigma2.sh.in] Add option for skin backdrop
This change adds the option for skins to include a backdrop.mvi image at the top level of the skin that will be used when that skin is selected.

The code first checks to see if a skin is defined in the "settings" file and uses it if found. If no skin is identified then the "skin.pyo" file is searched to see if a built-in skin has been defined. If found then that will be used. If no skin associated backdrop is identified then the code will progress to the original search in "/etc/enigma2/" then "/usr/share/" and finally look to use the "bootlogo.mvi" image.