FallbackTuner: Disable import epg from fallback tuner
It was a nice idea, but we cannot get it 100% stable.
While loading the epg,dat file other tasks could corrupt the cache and
make the box instable and in addition while saving the epg.dat on the
server box the created epg.dat could get corrupted.

I really need help so the epg.dat could be properly loaded and saved.

The issue could lead to wierd side effects.