[GraphMultiEPG.py] Fixed borders with solid background colors
This was actually a feature not implemented properly or implemented partially in the first place... Setting borders for the service and event rectangles was only possible when using background png's, but not when using sold background colors. The existing "ServiceBorderHorWidth", "ServiceBorderVerWidth", "EventBorderHorWidth" and "EventBorderVerWidth" skin attributes where only applied to background png's. This commit adds two new parameters, named just "ServiceBorderWidth" and "EventBorderHorWidth" to control the border width (same for horizontal and vertical dimensions) for the solid backgrounds rectangles.