[ServiceInfo.py] Re-introduced color for active ECM pid
This adds back the color option for singaling the active ECM pid. This was removed some time ago, because there was no option for the skin to control the color, which was hard coded into the code. Now, a new skin parameter is defined, named "ServiceInfoAltColor", that sets an alternative color to use in the ecm pid screen. Color values can be specified as decimal numbers, hex numbers, enigma2 "#" color codes or color names, like it is done e.g. in the virtual keyboard shift colors.