[skin.py] Experimental removal of SD and old skins
This commit is an experiment to test if the old SD default / OoB skin is stall actually being used or is still required.

This commit also tests the impact of removing the "skin_box.xml" and "skin_second_infobar.xml" skins.  It is believed that these skins are no longer used.  This test removal should prove the point.

If there are any issues for developers or users please immediately revert this commit and discuss the issues with me.