[SkinSelector.py] Adjust resolution search code
Change the way skin resolutions are obtained.  Now we will do an XML lookup for the resolution and if that fails we will revert to a text search.

This code is slower to execute than its predecessor but hopefully addresses Littlesat's reservations about the hack nature of the previous code.