cablescan: parse ServiceListDescriptors in the NIT
and ignore services which occur in the SDT but do not occur in the NIT
service list, for the current transponder.
It might happen that SDT data contains invalid information from the
original network, or from other network ids for instance. We should
assume the NIT contains the correct service list for each of the
occurring network ids, so everything else should be ignored.

One exception; when the NIT does not provide a service list, we keep
relying on the SDT like we did before.