Add a function to toggle subtitles on/off default with long subtitle
The toggle subtitles is default on long subtitle. It can also be added
as a long key function. I also added long subtitle as new key to Hotkey
to keep it in line.

todo: I used the VideoMode screen as information screen - we need to
rename this screen but it also need adaptions in skins so I decided to
sleep on this one.