Add support for auto resizable embeded skin templates
When the embeded skin included in python code is a list then this means
that this is an autoreziable skin template that can adjust to the
resulition of the skin. The first element in the list is the skin as xml
code where variable parameters are entered as %d and the following
elements are the values for an 1280x720 sized skin. During parsing the
embeded skin these values will be auto adjusted to the correct skin
resolution.

For an example see the next merge request (example in SkinSelecktor.py)

This avoids to call auto size code on multiple places as it now can be
done centralized.

I also choose not to use SkinTemplate and scaleData as two extra
variables for this as it can be done in one standard skin parameter due
to using the the check on list.