[UsageConfig.py] Transition screen path settings
This temporary code will find the user's current value for the old "config.usage.menu_path" setting and move it to the new "config.usage.showScreenPath".  The code will then remove the old "config.usage.menu_path" entry from the settings file.

This code is only run once such that when the setting is transferred the old setting is no longer used.  It is intended to remove this code after one or two production releases of OpenPLi when it will be assumed that most users have had the setting transferred.  When this code is removed the only implication for users will be that the setting will not be migrated and the "config.usage.showScreenPath" default value will be used.  Also the old "config.usage.menu_path" setting, if not at default, will not be removed.