[UsageConfig.py] Remove transition code
This change was requested by Littlesat and Persian Prince as OpenPLi production releases are too far apart such that this code could hang around for a very long time.  It is not considered an issue for this setting to be reset to default and the old value to be left in the setting file.