Show all available bouquets when leaving Wizard
When you exit the wizard you land on the userbouquet list. When during
the wizard services with new userbouquets were scanned and also when
import channels for fallback tuner was running and completed you do not
see the (new) bouqeuts. This commit ensures all the bouquets that are
added (scanned) are visible. It also hides the import channels popup
when during the usage of the wizard the import did complete.