filepush: part four.
We're struggling between "behaviour" of drivers from different
manufacturers. The ppoll and signal block works really well on
VU+, but on other brands it the code may block in other system
calls now (either the read() or the write() (somewhere else).

So let's see if we revert to the old behaviour without blocking
signals. Now every system call will be interrupted by a signal
(even poll). Combined with the better error checking of the result
of the call to poll, it may very well work now as expected.