filepush: make stopping recording thread more robust.
Keep sending the signal that's supposed to break the thread out of any
blocking syscall until the thread acknowledges that it finished. Only
then join() the thread. Join blocks until the thread is rally ended, so
if the thread keeps running or is blocked in a system call, the parent
will be blocked too, possibly forever.