filepush: improve Broadcom encoding workaround.
It appears read() does not return 0 in the case of no data, but it
returns -1 with errno = EAGAIN.

Normally this will never happen, as the file descriptors are never
opened O_NONBLOCK.

So when it occurs, call in the help from poll() and wait for actual data
to arrive.

This will not interfere with the normal path where data is fetched from
a demuxer or a file, this condition will not occur and poll will not
be involved.

(cherry picked from commit 957a9148255accc604a770ffc9acbbfa624fe4e3)