Do not show InfoBar on menu path
But I still have a wierd thing to solve. I have on my VU+Ultimo 4K now
that InfoBar is shown, but on my H9Combo it is not shown and now I miss
MainMenu and I have at this moment no idea why it is not shown. Maybe we
should consider to only show it from MainMenu so there is a kind of 2do
here....