[FastScan] Use named arguments for string
This fixes the following warning when updating the po files:

'msgid' format string with unnamed arguments cannot be properly localized:
The translator cannot reorder the arguments.
Please consider using a format string with named arguments,
and a mapping instead of a tuple for the arguments.

All existing translations are updated with the new string formatting.