Harddisk: Create "movie" directory after formatting a device (#3167)
Put your recordings into a convenient location, so you don't get to see
the "backup" and "epg.dat" entries. Easy enough to delete if you don't
want it, creating it is more work and not intuitive as the "movie" name
is sort of magic - the box will automatically use it when present.