Start enigma after build in workflow test
Now source files are compiled on test and this allows you to check for syntax errors.
But this does not allow check for example type or attribute errors, that may appear when an enigma run.
Therefore on test after the compile install and start enigma.
That will not check all errors in sources, but can prevent the possibility get gsod already on enigma startup.
The run doesn't happen on a real image, so a small surce patch is required for enigma startup.