Menu.py: Call setList to update __list
self.__list in the List.py module remains empty, because it's not set
with the constructor and setList is not being called.
Also the setter of the property() object 'list' is not being used (only getter).
So that make the codes regarding self.__list reduntant.

Copy code from OpenViX to try to fix that.

TODO: Test it and when OK, restore 'assert' code (in or without try/except block).