Long button press emulation
- added an ability to emulate long button presses with two normal
(short presses). One button on the remote must be defined for the long
press emulation purpose. Then a normal (short) press on the emulation
button followed by a normal (short) press on a second button emulates a
long press on the second button;
- new setting “Long press emulation with button XXX” on button setup
screen to define the button for that purpose.

-thanks openATV @team