Use existing renderer for gb led settings
In the enigma from the beginning there is a renderer FrontpanelLed which is used for led control.
Therefore, we can use it and do not need to additionally write values in the proc on startup.
This also introduces the ability to set led profiles for use in different modes for frontpanel led control.