MovieList: fix compare between str and int
On buildBeginTimeSortKey it was possible to set on second field the time on some entries
and on some other entries the name. Use default empty name and 0 for time to have consistent types.

This commit fixes the following problem:

Traceback (most recent call last):
  File "/usr/lib/enigma2/python/Screens/MovieSelection.py", line 1399, in reloadWithDelay
    self["list"].reload(self.current_ref, self.selected_tags)
  File "/usr/lib/enigma2/python/Components/MovieList.py", line 605, in reload
    self.load(root, filter_tags)
  File "/usr/lib/enigma2/python/Components/MovieList.py", line 715, in load
    self.list.sort(key=self.buildGroupwiseSortkey)
TypeError: '<' not supported between instances of 'str' and 'int'