Introduces partial event search in EventView and EpgSelection
These screens already have a similar event search, but it does not show all possible events.
Therefore, the option to search for partial events is added to the unused yellow button, so that partially similar events can also be found and used.

Also remove imported but unused and fix over-indented.