Move hotplug ipk files from SoftwareManager to Hotplug plugin
As the software manager is not installed by default as it is not really
needed anymore. However it still has the hotplugin for ipk files. By
moving the hotplut for ipk files to the Hotplug plugin, which is
installed by default, the insert USB stick with ipk file will always
work by default in OpenPli