Add propery in config for ConfigSelectionNumber
It seems this was a long time forgotton property that results the config
will return a string instead of an interger. But since python 3 this
might go wrong (see the crash log example below) within for example the
autovolume plugin. I tested it for a couple of hours and until now I
could not find any side effects.

Traceback (most recent call last):
  File "/usr/lib/enigma2/python/Navigation.py", line 86, in dispatchEvent
  File "/usr/lib/enigma2/python/Components/ServiceEventTracker.py", line 82, in event
  File "/usr/lib/enigma2/python/Plugins/SystemPlugins/AutomaticVolumeAdjustment/AutomaticVolumeAdjustment.py", line 128, in __evUpdatedInfo
TypeError: '>' not supported between instances of 'int' and 'str'
[ePyObject] (CallObject(<bound method Navigation.dispatchEvent of <Navigation.Navigation object at 0xb212d898>>,(5,)) failed)