Restore title for MovieSelection to as it was before
For a specific new coming up skin in the title of the movie selection
the free disk space was shown. This was in most standard skin a specific
widget. Unless this the feature is still nice to have it but we think it
should be adaptable e.g. via the parameter section of a skin. So the
standard existing skins can use what is currently available and when a
new skin needs this special 'combined' feature then do it via a
parameter or something else