Remove the immutable work-a-rounds from BoxInfo.
As actually it does not have a real added value as I also discussed with
Huevos. Read e.g. the discussion here: https://forums.openpli.org/topic/97296-boxinfo-discussion/page-5
It is possible to add complicated code but at the end with
BoxInfo.boxInfo = {} you can clear all and/or overwrite everything. You
only try to protect when someone with coding does something by accident.
By not checking this the code looks very lean and both worlds SystemInfo
and BoxInfo work and are fully in Sync. So when someone changed a key in
SystemInfo, BoxInfo gives the same result and visa versa.