Fix split on only the first = in enigma.info and strip differently
Somewhere in this adventure I lost to split it only on the first =
The orriginal implementation did strip the item and the value in a loop.
For openpli's enigma.info only to strip on the complete line is
sufficiant but it seems other distributions have at the value leading
spaces. Actually the strip is only needed for the value. But to stay
compatible when someone has an idea to add leading and trailing spaces
to the item I considered to strip the item as well. But not in the loop.
Acutally this is only needed for parsing other image's enigma.info file
(for Multiboot purposes).

For multiboot purposes it would be helpfull to get an item added that
has a value how the image wants to show up in the multiboot selection