For repeated timers do not take the middle for the EIT.
This is not fail save, but a quick improvement. Usually you will have at the beginning at least the half or more from the movie. Therefor most likely when you take the sample not at the middle but at eg 20% or so you get a better hit. I know it is not and never 100% but it might give us some time to think about something better. In the situation with openwebif sync the epg is one option but even then you cannot guarantee and here also this idea will improve it. But here you can also check the description.
Taking the begin time and add the marge before recording might be a good
idea but sometimes people might extend this manually as additional
safetumargin so this is also not a good 'trick'.