Make in BoxInfo and SystemInfo the data from enigma.info more immutable
It is not fail safe but we prefer to have the data that is coming from
enigma.info at least to have better protected so the data is telling the
truth. So when someone tries to change such value that it rises a
TypeError. We know someone who will really work-a-round it can still
harm it, but it at least protects from someone who is doing it
unintended. Together we ensure that the legacy SystemInfo dict has the
same and is in sync with the BoxInfo.boxInfo dict.