Revert "Implement getEnigmaConfList so when needed it can be used"
This reverts commit 77ee3b2e78ddcb70492865142eba575067144cd4.

It seems to be needed for OpenATV that can have an alternative
enigma.conf list to overrult the enigma.info file and this plugin get a
list of items that were included in this alternative list. As we do not
have this kind of bypass the change I mande here was not what I thouht
it was for. So we can indeed better return an empty list