A config conditionally hidden but changed should be saved
In the previous work-a-round I did save all when the config was rebuild
but of course we only need to save configs that are conditionally not
shown only when they are changed.

I get to the config to an eval at this moment as I could not (quickly)
find a better way to get to the specific config

This avoids e.g. in the auto language setup when we put languages on
'None' while they disappear from the config list that the value is
actually not saved

See:
https://forums.openpli.org/topic/98887-auto-language-selection-options-not-saved-corectly/